# Conflicts:
#	src/main/java/com/finnegans/gestioncrisalis/dtos/EmpresaDTO.java
#	src/main/java/com/finnegans/gestioncrisalis/services/impl/EmpresaServiceImpl.java
#	src/main/java/com/finnegans/gestioncrisalis/validations/EmpresaDTOValidator.java

+++ src/main/java/com/finnegans/gestioncrisalis/dtos/request/EmpresaResponseDTO.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class EmpresaResponseDTO {
     @JsonProperty("id")
-    private int idDTO;
+    private Long idDTO;
 
     @JsonProperty("nombre")
     private String nombreDTO;
